nav component split to clear navigation logic
