+++ frontend/src/components/navigation/index.js
@@ -1,45 +1,39 @@
 import './index.css';
 import logo from './logo.png';
-import { MdLogout, MdPayment } from "react-icons/md";
 import { FiUsers } from "react-icons/fi";
 import { GrUserAdmin } from "react-icons/gr";
-import { IoArrowBackOutline, IoStatsChart } from "react-icons/io5";
-import { IoIosArrowDown } from "react-icons/io";
-import { RxHamburgerMenu, RxReader } from "react-icons/rx";
+import { MdPayment } from "react-icons/md";
+import { RxReader } from "react-icons/rx";
+import { IoStatsChart } from "react-icons/io5";
+import { IoArrowBackOutline } from "react-icons/io5";
+import { RxHamburgerMenu } from "react-icons/rx";
 import { useState, useEffect } from 'react';
 import { Link, useLocation } from "react-router-dom";
 import { useTranslation } from 'react-i18next';
-import { useAuth } from '../../contexts/auth-context';
-import LanguageSwitcher from '../language-switcher';
-import Modal from '../modal';
+import UserPanel from '../user-panel';
 import api from '../../configs/api';
-import { useForm } from "react-hook-form";
-import { FiEye, FiEyeOff } from "react-icons/fi";
 
 const Navigation = ({ changeLanguage }) => {
-  const { t, i18n } = useTranslation();
-  const { logout, userData } = useAuth();
+  const { t } = useTranslation();
   const [loading, setLoading] = useState(true);
   const [userEmail, setUserEmail] = useState(null);
   const location = useLocation();
   const userId = location.pathname.split('/')[3];
-  const [isAccordionOpen, setIsAccordionOpen] = useState(false);
   const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [notification, setNotification] = useState({ isOpen: false, message: "" });
-  const [showOldPassword, setShowOldPassword] = useState(false);
-  const [showNewPassword, setShowNewPassword] = useState(false);
 
-  const { register, handleSubmit, formState: { errors }, reset } = useForm();
+  const navLinks = [
+    { path: '/customers', icon: <FiUsers className='icon' />, label: t('navigation.customers') },
+    { path: '/admins', icon: <GrUserAdmin className='icon' />, label: t('navigation.admins') },
+    { path: '/subscriptions', icon: <MdPayment className='icon' />, label: t('navigation.subscriptions') },
+    { path: '/audit-log', icon: <RxReader className='icon' />, label: t('navigation.auditLog') },
+    { path: '/statistics', icon: <IoStatsChart className='icon' />, label: t('navigation.statistics') },
+  ];
 
   useEffect(() => {
-    if (i18n.isInitialized) {
+    if (t('navigation.customers')) {
       setLoading(false);
-    } else {
-      const handleInitialized = () => setLoading(false);
-      i18n.on('initialized', handleInitialized);
-      return () => i18n.off('initialized', handleInitialized);
     }
-  }, [i18n]);
+  }, [t]);
 
   useEffect(() => {
     if (!location.pathname.startsWith("/customers/user")) {
@@ -63,34 +57,8 @@ const Navigation = ({ changeLanguage }) => {
     }
   }, [userId]);
 
-  const handleLogout = () => {
-    logout();
-  };
-
-  const toggleAccordion = () => {
-    setIsAccordionOpen(!isAccordionOpen);
-  };
-
   const toggleMenu = () => {
     setIsMenuOpen(!isMenuOpen);
-  };
-
-  const onSubmit = async (data) => {
-    try {
-      const response = await api.put('/user/password', {
-        old_password: data.oldPassword,
-        new_password: data.newPassword,
-      });
-      if (response.status === 200) {
-        setNotification({ isOpen: true, message: t("notification.passwordUpdated") });
-        setTimeout(() => {
-          toggleAccordion();
-          reset();
-        }, 700);
-      }
-    } catch (error) {
-      setNotification({ isOpen: true, message: t("notification.passwordUpdateFailed") });
-    }
   };
 
   const getActiveLink = (path) => {
@@ -129,123 +97,31 @@ const Navigation = ({ changeLanguage }) => {
       </div>
       <div className='navigation-container'>
         <div className='links'>
-          <Link
-            to="/customers"
-            className={`link ${getActiveLink('/customers')} ${userEmail ? 'email' : ''}`}
-            onClick={handleLinkClick}
-          >
-            <div className='customer-link'>
-              {!userEmail ? (
-                <FiUsers className='icon' />
-              ) : (
-                <IoArrowBackOutline className='icon arrow' />
+          {navLinks.map(({ path, icon, label }) => (
+            <Link
+              key={path}
+              to={path}
+              className={`link ${getActiveLink(path)} ${userEmail && path === '/customers' ? 'email' : ''}`}
+              onClick={handleLinkClick}
+            >
+              <div className='customer-link'>
+                {userEmail && path === '/customers' ? (
+                  <IoArrowBackOutline className='icon arrow' />
+                ) : (
+                  icon
+                )}
+                {label}
+              </div>
+              {userEmail && path === '/customers' && (
+                <div className="user-email">
+                  {userEmail}
+                </div>
               )}
-              {t('navigation.customers')}
-            </div>
-            {userEmail && (
-              <div className="user-email">
-                {userEmail}
-              </div>
-            )}
-          </Link>
-          <Link
-            to="/admins"
-            className={`link ${getActiveLink('/admins')}`}
-            onClick={handleLinkClick}
-          >
-            <GrUserAdmin className='icon' />
-            {t('navigation.admins')}
-          </Link>
-          <Link
-            to="/subscriptions"
-            className={`link ${getActiveLink('/subscriptions')}`}
-            onClick={handleLinkClick}
-          >
-            <MdPayment className='icon' />
-            {t('navigation.subscriptions')}
-          </Link>
-          <Link
-            to="/audit-log"
-            className={`link ${getActiveLink('/audit-log')}`}
-            onClick={handleLinkClick}
-          >
-            <RxReader className='icon' />
-            {t('navigation.auditLog')}
-          </Link>
-          <Link
-            to="/statistics"
-            className={`link ${getActiveLink('/statistics')}`}
-            onClick={handleLinkClick}
-          >
-            <IoStatsChart className='icon' />
-            {t('navigation.statistics')}
-          </Link>
+            </Link>
+          ))}
         </div>
-        <div className='user-panel'>
-          <div className='user-header'>
-            <p className='name'>{userData?.name || ''}</p>
-            <p className='email'>{userData?.email || ''}</p>
-            <IoIosArrowDown
-              className={`icon ${isAccordionOpen ? 'open' : ''}`}
-              onClick={toggleAccordion}
-            />
-            {isAccordionOpen && (
-              <div className='accordion-content'>
-                <form className='form' onSubmit={handleSubmit(onSubmit)}>
-                  <div className='form-group'>
-                    <input
-                      type={showOldPassword ? 'text' : 'password'}
-                      {...register('oldPassword', {
-                        required: t('login.passwordRequired'),
-                        minLength: { value: 8, message: t('login.passwordMinLength', { length: 8 }) },
-                      })}
-                      placeholder={t('navigation.oldPassword')}
-                    />
-                    <button
-                      type="button"
-                      className="icon"
-                      onClick={() => setShowOldPassword(!showOldPassword)}
-                    >
-                      {showOldPassword ? <FiEyeOff /> : <FiEye />}
-                    </button>
-                  </div>
-                  {errors.oldPassword && <p className="error">{errors.oldPassword.message}</p>}
-                  <div className='form-group'>
-                    <input
-                      type={showNewPassword ? 'text' : 'password'}
-                      {...register('newPassword', {
-                        required: t('login.passwordRequired'),
-                        minLength: { value: 8, message: t('login.passwordMinLength', { length: 8 }) },
-                      })}
-                      placeholder={t('navigation.newPassword')}
-                    />
-                    <button
-                      type="button"
-                      className="icon"
-                      onClick={() => setShowNewPassword(!showNewPassword)}
-                    >
-                      {showNewPassword ? <FiEyeOff /> : <FiEye />}
-                    </button>
-                  </div>
-                  {errors.newPassword && <p className="error">{errors.newPassword.message}</p>}
-                  <button type='submit'>{t('navigation.changePassword')}</button>
-                </form>
-                <LanguageSwitcher changeLanguage={changeLanguage} />
-              </div>
-            )}
-          </div>
-          <button className='logout' onClick={handleLogout}>
-            <MdLogout className='icon' />
-            <span>{t('navigation.logout')}</span>
-          </button>
-        </div>
+        <UserPanel changeLanguage={changeLanguage} />
       </div>
-      <Modal
-        isOpen={notification.isOpen}
-        onClose={() => setNotification({ isOpen: false, message: "" })}
-        message={notification.message}
-        showCloseButton={true}
-      />
     </div>
   );
 };
