+++ frontend/src/components/user-panel/index.css
@@ -1,0 +1,344 @@
+.navigation {
+  background: white;
+  margin: 2rem;
+  border-radius: 1rem;
+  padding: 2rem;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  min-width: fit-content;
+  height: calc(100vh - 4rem);
+  box-sizing: border-box;
+}
+
+.navigation .navigation-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+}
+
+.navigation .logo {
+  display: flex;
+  justify-content: center;
+  font-family: Kreon;
+  font-size: 2rem;
+  user-select: none;
+  position: relative;
+}
+
+.navigation .logo .desktop-logo {
+  width: 48px;
+  height: 40px;
+  padding-right: 1rem;
+}
+
+.navigation .logo .desktop-logo,
+.desktop-text {
+  display: block;
+}
+
+.navigation .logo .mobile-title {
+  display: none;
+}
+
+.navigation .logo span {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #333;
+}
+
+.navigation .links {
+  margin-top: 2rem;
+}
+
+.navigation .link {
+  font-size: 1.5rem;
+  display: flex;
+  align-items: center;
+  color: #333;
+  text-decoration: none;
+  margin-bottom: 1rem;
+  padding: 0.5rem 1rem;
+  border-radius: 0.5rem;
+  transition: background 0.3s;
+}
+
+.navigation .link.email {
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.navigation .link.active {
+  background: #dedede;
+}
+
+.navigation .link:hover {
+  background: #e6e6e6;
+}
+
+.navigation .link .icon {
+  margin-right: 1rem;
+  font-size: 1.8rem;
+}
+
+.navigation .link .customer-link {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.navigation .link .icon.arrow {
+  font-size: 2rem;
+}
+
+.navigation .link .user-email {
+  font-size: 1.1rem;
+  color: #333;
+  padding-top: 0.5rem;
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.navigation .link .user-email .icon {
+  margin: 0;
+}
+
+.navigation .navigation-container .user-panel {
+  margin-top: auto;
+  position: relative;
+}
+
+.navigation .user-panel .user-header {
+  background: #D5D5D5;
+  padding: 0.5rem 1rem;
+  padding-top: 0;
+  border-radius: 0.5rem;
+  min-height: 2rem;
+}
+
+.navigation .user-panel .user-header .icon {
+  font-size: 1.5rem;
+  transition: transform 0.3s;
+  cursor: pointer;
+  position: absolute;
+  top: 0.5rem;
+  right: 0.8rem;
+  color: #333;
+}
+
+.navigation .user-panel .user-header .icon.open {
+  transform: rotate(180deg);
+}
+
+.navigation .user-header p {
+  color: #333;
+  margin: 0;
+}
+
+.navigation .user-panel .user-header .name {
+  font-size: 1.7rem;
+}
+
+.navigation .user-panel .user-header .email {
+  font-size: 1.2rem;
+}
+
+.navigation .user-panel .logout {
+  width: 100%;
+  margin-top: 1rem;
+  border-radius: 0.5rem;
+}
+
+.navigation .user-panel .user-header .accordion-content .form {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  margin-top: 1rem;
+}
+
+.navigation .user-panel .user-header .accordion-content .form-group {
+  margin-bottom: 0.5rem;
+  font-size: 1rem;
+  padding-left: 1rem;
+  background: white;
+  border-radius: 0.5rem;
+  position: relative;
+}
+
+.navigation .user-panel .user-header .accordion-content .form-group .icon {
+  top: 50%;
+  right: 0.7rem;
+  transform: translateY(-50%);
+  padding: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.navigation .user-panel .user-header .accordion-content input,
+.navigation .user-panel .user-header .accordion-content button[type="submit"] {
+  outline: none;
+  border: none;
+  color: #333;
+  border-radius: 0.5rem;
+  height: 2rem;
+}
+
+.navigation .user-panel .user-header .accordion-content button[type="submit"] {
+  margin-bottom: 1rem;
+  background: #222;
+  color: white;
+  font-size: 1rem;
+}
+
+.navigation .user-panel .user-header .accordion-content input {
+  padding: 0;
+}
+
+.navigation .user-panel .user-header .accordion-content input::placeholder {
+  font-size: 1rem;
+}
+
+.navigation .user-panel .user-header .accordion-content .error {
+  padding: 0.5rem;
+  padding-top: 0;
+}
+
+.navigation .user-panel .logout {
+  height: 3rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: .3s;
+  background: #FF8585;
+}
+
+.navigation .user-panel .logout:hover {
+  background: #fd5656;
+}
+
+.navigation .user-panel .logout .icon {
+  height: 2rem;
+  width: 2rem;
+  color: white;
+}
+
+.navigation .user-panel .logout span {
+  font-size: 1.5rem;
+  color: white;
+  user-select: none;
+  padding-left: 0.5rem;
+  width: 30%;
+}
+
+.navigation.expanded {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: calc(100% - 2rem);
+  height: calc(100vh - 2rem);
+  z-index: 1000;
+}
+
+.burger-menu {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: white;
+  z-index: 1000;
+  display: flex;
+  flex-direction: column;
+  padding: 1rem;
+}
+
+.burger-menu .menu-links {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.burger-menu .logout {
+  margin-top: auto;
+  background: #ff8585;
+  color: white;
+  padding: 0.5rem;
+  border: none;
+  border-radius: 0.5rem;
+  cursor: pointer;
+}
+
+.burger-menu .logout:hover {
+  background: #ff5656;
+}
+
+.burger-menu-icon {
+  font-size: 2rem;
+  cursor: pointer;
+  display: none;
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  left: 0;
+}
+
+@media (max-width: 768px) {
+
+  .navigation .logo .desktop-logo,
+  .navigation .logo .desktop-text {
+    display: none;
+  }
+
+  .navigation .logo .mobile-title {
+    display: block;
+    font-size: 1.4rem;
+    color: #333;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+      sans-serif;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    max-width: 70%;
+    text-align: center;
+  }
+
+  .navigation .navigation-container {
+    display: none;
+  }
+
+  .navigation.expanded .navigation-container {
+    display: flex;
+  }
+
+  .navigation {
+    padding: 1rem;
+    height: fit-content;
+    width: calc(100% - 2rem);
+    margin: 1rem;
+    margin-bottom: 0;
+  }
+
+  .burger-menu-icon {
+    display: block;
+  }
+
+  .navigation .logo {
+    font-size: 1.2rem;
+  }
+
+  .navigation .logo img {
+    width: 48px;
+    height: 40px;
+    padding-right: 1rem;
+  }
+
+  .burger-menu-icon {
+    display: block;
+  }
+}
